+++ src/lib/scheduler_clients/__init__.py
@@ -4,15 +4,10 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 # models
-from lib.scheduler_clients.slurm.models import (
-    SlurmJobDescription,
-    SlurmJob,
-    SlurmNode,
-)
-from lib.scheduler_clients.pbs.models import (
-    PbsJob,
-    PbsJobDescription,
-    PbsNode,
+from lib.scheduler_clients.models import (
+    JobModel,
+    JobDescriptionModel,
+    NodeModel,
 )
 
 # clients
@@ -22,11 +17,7 @@ from lib.scheduler_clients.slurm.slurm_rest_client import (
 
 
 __all__ = [
-    "PbsJob",
-    "PbsJobDescription",
-    "PbsNode",
-    "SlurmJob",
-    "SlurmJobDescription",
-    "SlurmNode",
-    "SlurmRestClient",
+    "JobModel",
+    "JobDescriptionModel",
+    "NodeModel",
 ]
